Start work on chapter 2.
This is a big-ish commit. It contains some initial text/notes for
Chapter 2, as well as updates to the main diss file (adding ch. 2 as a
subfile), the bib file, and a graphics subproject update. This commit
also adds a \caph command, to set harmonies in captions using the
sans-serif caption font.