Don't use subfiles package.
By using LaTeX's include/includeonly, it's possible to generate the aux
files for individual chapters (by running xelatex on the entire
document), and then work on individual chapters (using \includeonly).
This means that later chapters can look at the .aux files for the
earlier chapters to get the page/figure/example numbers correct. That
way when you're working on chapter 2, it actually says chapter 2 at the
top of the page.

(This is a stupid thing to be complaining about, but it's annoying to
me.)