Change mathematical labels to be more descriptive.
This changes S_1, S_2, and S_3 to be S_min, S_dom, and S_maj,
respectively. It also changes the generic x_1, x_2, and x_3 to be x_r,
x_t, and x_s (root, third, and seventh). This makes the math a little
easier to follow throught and more meaningful.