Change the "7" in slide7 to a subscript.
I did this because in the next section we'll need the inverse of slide7
a lot, and (slide^7)^-1 looks silly. It's also misleading: it has the
potential to be read as slide^(-7). The seven was never an exponent;
it's just a label to differentiate the seventh-chord slide from the
triadic (Lewinnian) slide, so a subscript makes more sense.

NB! This necessitates a change in all of the graphics which use slide7,
which I haven't done yet.